# examples/OpenAPI_examples/SmartHome/SmartCity_Homes/SmartHome1_OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartHome1",
    "description": "This is a Web Thing Model server.  You can find\nout more about Web Thing Model (W3C) at\n[https://www.w3.org/Submission/wot-model/](https://www.w3.org/Submission/wot-model/).\n",
    "x-location" : "Chania",
    "contact": {
      "email": "atzavaras@intelligence.tuc.gr"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://172.16.1.1:5002/SmartHome_actuator1",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
...
    "/properties": {
      "get": {
        "tags": [
          "Properties"
        ],
        "summary": "Retrieve a list of properties",
        "description": "In response to an HTTP GET request on the destination URL of a properties link, an Extended Web Thing must return an array of Property that the initial resource contains.",
        "operationId": "retrieveWebThingProperties",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertiesResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/actions": {
      "get": {
        "tags": [
          "Actions"
        ],
        "summary": "Retrieve a list of actions",
        "operationId": "retrieveWebThingActions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Action"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PropertiesResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TempProperty"
          },
          {
            "$ref": "#/components/schemas/HumProperty"
          }
        ],
        "xml": {
          "name": "PropertiesResponse"
        }
      },
      "TempMeasurement": {
        "type": "object",
        "x-kindOf": "http://www.w3.org/ns/sosa/Observation",
        "properties": {
          "temp": {
            "type": "integer",
            "example": 25,
            "x-kindOf": "http://www.w3.org/ns/sosa/hasSimpleResult"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "x-kindOf": "http://www.w3.org/ns/sosa/resultTime"
          }
        }
      },
      "HMeasurement": {
        "type": "object",
        "x-kindOf": "http://www.w3.org/ns/sosa/Observation",
        "properties": {
          "h": {
            "type": "integer",
            "example": 21,
            "x-kindOf": "http://www.w3.org/ns/sosa/hasSimpleResult"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "x-kindOf": "http://www.w3.org/ns/sosa/resultTime"
          }
        }
      },
      "TempProperty": {
        "type": "object",
        "required": [
          "id",
          "values"
        ],
        "x-kindOf": "http://www.w3.org/ns/ssn/systems/SystemCapability",
        "properties": {
          "id": {
            "type": "string",
            "default": "DHT22_temperature",
            "x-kindOf": "http://schema.org/identifier"
          },
          "name": {
            "type": "string",
            "example": "Temperature",
            "x-kindOf": "http://schema.org/name"
          },
          "values": {
            "$ref": "#/components/schemas/TempMeasurement"
          },
          "range": {
            "type": "object",
            "x-kindOf": "http://www.w3.org/ns/ssn/systems/Condition",
            "properties": {
              "unit": {
                "type": "string",
                "default": "DegreeCelsius",
                "x-refersTo": "https://schema.org/unitText"
              },
              "minValue": {
                "type": "number",
                "format": "float",
                "default": -40,
                "x-refersTo": "https://schema.org/value"
              },
              "maxValue": {
                "type": "number",
                "format": "float",
                "default": 80,
                "x-refersTo": "https://schema.org/value"
              }
            }
          },
          "sensorAccuracy": {
            "type": "object",
            "x-refersTo": "http://www.w3.org/ns/ssn/systems/Accuracy",
            "properties": {
              "range": {
                "type": "object",
                "x-kindOf": "http://www.w3.org/ns/ssn/systems/Condition",
                "properties": {
                  "unit": {
                    "type": "string",
                    "default": "DegreeCelsius",
                    "x-refersTo": "https://schema.org/unitText"
                  },
                  "minValue": {
                    "type": "number",
                    "format": "float",
                    "default": -0.5,
                    "x-refersTo": "https://schema.org/value"
                  },
                  "maxValue": {
                    "type": "number",
                    "format": "float",
                    "default": 0.5,
                    "x-refersTo": "https://schema.org/value"
                  }
                }
              }
            }
          },
          "sensorSensitivity": {
            "type": "object",
            "x-refersTo": "http://www.w3.org/ns/ssn/systems/Sensitivity",
            "properties": {
              "unit": {
                "type": "string",
                "default": "DegreeCelsius",
                "x-refersTo": "https://schema.org/unitText"
              },
              "value": {
                "type": "number",
                "format": "float",
                "default": 0.1,
                "x-refersTo": "https://schema.org/value"
              }
            }
          },
          "sensorPrecision": {
            "type": "object",
            "x-refersTo": "http://www.w3.org/ns/ssn/systems/Precision",
            "properties": {
              "range": {
                "type": "object",
                "x-kindOf": "http://www.w3.org/ns/ssn/systems/Condition",
                "properties": {
                  "unit": {
                    "type": "string",
                    "default": "DegreeCelsius",
                    "x-refersTo": "https://schema.org/unitText"
                  },
                  "minValue": {
                    "type": "number",
                    "format": "float",
                    "default": 0.2,
                    "x-refersTo": "https://schema.org/value"
                  },
                  "maxValue": {
                    "type": "number",
                    "format": "float",
                    "default": 0.2,
                    "x-refersTo": "https://schema.org/value"
                  }
                }
              }
            }
          },
          "sensorFrequency": {
            "type": "object",
            "x-refersTo": "http://www.w3.org/ns/ssn/systems/Frequency",
            "properties": {
              "unit": {
                "type": "string",
                "default": "sec",
                "x-refersTo": "https://schema.org/unitText"
              },
              "period": {
                "type": "integer",
                "format": "int32",
                "default": 2,
                "x-refersTo": "https://schema.org/value"
              }
            }
          }
        },
        "xml": {
          "name": "TempProperty"
        }
      },
      "HumProperty": {
        "type": "object",
        "required": [
          "id",
          "values"
        ],
        "x-kindOf": "http://www.w3.org/ns/ssn/systems/SystemCapability",
        "properties": {
          "id": {
            "type": "string",
            "default": "DHT22_humidity",
            "x-kindOf": "http://schema.org/identifier"
          },
          "name": {
            "type": "string",
            "example": "Humidity",
            "x-kindOf": "http://schema.org/name"
          },
          "values": {
            "$ref": "#/components/schemas/HMeasurement"
          },
          "range": {
            "type": "object",
            "x-kindOf": "http://www.w3.org/ns/ssn/systems/Condition",
            "properties": {
              "unit": {
                "type": "string",
                "default": "Percent",
                "x-refersTo": "https://schema.org/unitText"
              },
              "minValue": {
                "type": "number",
                "format": "float",
                "default": 5,
                "x-refersTo": "https://schema.org/value"
              },
              "maxValue": {
                "type": "number",
                "format": "float",
                "default": 85,
                "x-refersTo": "https://schema.org/value"
              }
            }
          }
        },
        "xml": {
          "name": "ΗumProperty"
        }
      },
      "Action": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "on"
          },
          "name": {
            "type": "string",
            "example": "Switch on the Smart Lamp"
          }
        },
        "xml": {
          "name": "Action"
        }
      },
...
    },
...
  }
}